Move liveshare into datascience
