+++ src/test/datascience/mockLiveShare.ts
@@ -7,10 +7,10 @@ import { CancellationToken, CancellationTokenSource, Disposable, Event, EventEmi
 import * as vsls from 'vsls/vscode';
 
 import { IApplicationShell, ILiveShareTestingApi } from '../../client/common/application/types';
-import { LiveShareProxy } from '../../client/common/liveshare/liveshareProxy';
 import { IConfigurationService, IDisposable, IDisposableRegistry } from '../../client/common/types';
 import { noop } from '../../client/common/utils/misc';
 import { LiveShare } from '../../client/datascience/constants';
+import { LiveShareProxy } from '../../client/datascience/liveshare/liveshareProxy';
 
 // tslint:disable:no-any unified-signatures max-classes-per-file
 
@@ -355,7 +355,7 @@ class MockLiveShare implements vsls.LiveShare, vsls.Session, vsls.Peer, IDisposa
         return Promise.resolve({ dispose: noop });
     }
 
-    private generateServicePair() : MockLiveService[] {
+    private generateServicePair(): MockLiveService[] {
         const hostService = new MockLiveService();
         const guestService = new MockLiveService();
         hostService.setSibling(guestService);
@@ -375,9 +375,9 @@ export class MockLiveShareApi implements ILiveShareTestingApi {
 
     constructor(
         @inject(IDisposableRegistry) private disposables: IDisposableRegistry,
-        @inject(IApplicationShell) private appShell : IApplicationShell,
+        @inject(IApplicationShell) private appShell: IApplicationShell,
         @inject(IConfigurationService) private config: IConfigurationService
-        ) {
+    ) {
     }
 
     public getApi(): Promise<vsls.LiveShare | null> {
